# Skip check for existing files when reading file data from local DB