As discussed in #10783 and #11181.

It seems that the code to check for existing files when reading file data from the local DB is no longer needed.

In my "standard" benchmark (~330 files in a directory on a low-powered device), with this change  the average execution time of the `OCFileListFragment::listDirectory` call is reduced from **~320 ms** to **~280 ms** (but the timings are not very consistent).

I have not noticed any negative side effects, but to be honest, I have no idea how to really test it.