@starypatyk 2023-01-06
One of the test failures seems to be directly related to the changes:
[com.owncloud.android.FileIT - testRenameFolder](https://www.kaminsky.me/nc-dev/android-integrationTests/7270-IT-stable-23-21/com.owncloud.android.FileIT.html)

I have tried to repeat the issue in the app (renaming folder manually), but this works fine for me.

I am not sure yet, if the test assumes too much or if the storagePath should be set somewhere to compensate for the removed code.

Other failures in [EndToEndRandomIT](https://www.kaminsky.me/nc-dev/android-integrationTests/7270-IT-stable-23-21/com.nextcloud.client.EndToEndRandomIT.html) seem unrelated - at least I do not see any connection.

On the other hand the failure in [DialogFragmentIT](https://www.kaminsky.me/nc-dev/android-integrationTests/11227-Screenshot-blue-Light-21-35/com.owncloud.android.ui.dialog.DialogFragmentIT.html#testAccountChooserDialog) might be somehow related. It might be caused by a change of the timing between test execution in `DialogFragmentIT::testAccountChooserDialog` and delayed execution in the `FileDisplayActivity::startSyncFolderOperation`.