@starypatyk 2023-01-06
It turned out that the `file.exists` check is still required for directories.

When a folder is synced locally `storagePath` is set for files, but not for directories.

With the original change the **On Device** pane stopped showing folders. :-1: 

I had to partially revert the previous commit and leave the check for directories. I hope this will also fix the failing **testRenameFolder** test.

Assuming that in the typical situation there are much more files than folders, this still gives similar performance improvement as the first attempt.

BTW - I have found one reason why I get inconsistent timings. At some point after starting the app, the JIT compiler decides to optimize the `FileDao_Impl.getFolderContent` method:

`Compiler allocated 4MB to compile java.util.List com.nextcloud.client.database.dao.FileDao_Impl.getFolderContent(long)
`

This influences the results significantly. Previously I had not paid attention to this message, so I got skewed results depending on the moment the method got optimized. :disappointed: