@AlvaroBrey 2023-01-07
> At some point after starting the app, the JIT compiler decides to optimize the `FileDao_Impl.getFolderContent` method

If this kind of thing interests you, you may want to look into [Baseline profiles](https://developer.android.com/topic/performance/baselineprofiles/overview) so that critical methods like this one get compiled at install time.