@starypatyk 2023-01-07
> If this kind of thing interests you, you may want to look into [Baseline profiles](https://developer.android.com/topic/performance/baselineprofiles/overview) so that critical methods like this one get compiled at install time.

Interesting indeed. :wink: For now, there are still a few potential "regular" improvements in the pipeline. Would be good to learn this at some point.

> I had to partially revert the previous commit and leave the check for directories. I hope this will also fix the failing testRenameFolder test.

If I see correctly the **testRenameFolder** test is passing now. But I am struggling to find the results...