@starypatyk 2023-01-07
I have repeated the baseline tests and tests with the changes from this PR - this time making sure that in all cases the tests are executed **after** JIT decides to optimize the `FileDao_Impl.getFolderContent` method.

Method | Baseline - avg ms | PR - avg ms
-- | -- | --
OCFileListFragment::listDirectory | 291 | 234
FileDataStorageManager::getFolderContent | 168 | 112

We are gaining **~55 ms** when skipping ~330 `file.exists` calls.