@AlvaroBrey 2023-01-09
> When a folder is synced locally `storagePath` is set for files, but not for directories

I wonder why it's computed but not saved. @tobiasKaminsky any insights here?

Other than that, this is :+1:  for me