@AlvaroBrey 2023-01-23
OK to merge this for now, thanks @starypatyk . I still don't know why the path is not stored for folders, but we'll go ahead with the current check.